# docs/openapi.yaml
openapi: 3.0.3
info:
  title: RA Devs JWT Auth API
  description: |
    Universal JWT authentication API for Laravel applications.

    ## Features
    - Login / Logout / Register
    - Token Refresh via HTTP-only cookies
    - Password Reset via 8-character alphanumeric codes
    - Rate limiting on all endpoints
    - Security event logging

    ## Authentication
    Most endpoints require a Bearer token in the Authorization header:
    ```
    Authorization: Bearer {access_token}
    ```
  version: 1.0.0
  contact:
    name: RA Devs
    email: m.yastremskyj@gmail.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost/api/auth
    description: Local development server
  - url: https://api.example.com/api/auth
    description: Production server
...
paths:
...
  /refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      description: Generate a new access token using the refresh token from cookie.
      operationId: refresh
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '403':
          $ref: '#/components/responses/InvalidToken'
        '429':
          $ref: '#/components/responses/TooManyRequests'
...
  /reset-password:
    post:
      tags:
        - Password Reset
      summary: Reset password
      description: Reset user password using the reset code sent via email.
      operationId: resetPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '200':
          description: Password reset successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/responses/InvalidResetCode'
        '404':
          $ref: '#/components/responses/UserNotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: refresh_token
  schemas:
...
    ResetPasswordRequest:
      type: object
      required:
        - email
        - code
        - password
        - password_confirmation
      properties:
        email:
          type: string
          format: email
          example: john@example.com
        code:
          type: string
          minLength: 8
          maxLength: 8
          pattern: '^[A-Z2-9]{8}$'
          example: ABCD1234
        password:
          type: string
          format: password
          minLength: 8
          example: NewPassword123!
        password_confirmation:
          type: string
          format: password
          example: NewPassword123!
    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: John
        last_name:
          type: string
          example: Doe
        email:
          type: string
          format: email
          example: john@example.com
    Token:
      type: object
      properties:
        access_token:
          type: string
          example: eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9...
        token_type:
          type: string
          example: bearer
        expires_in:
          type: integer
          description: Token expiration time in seconds
          example: 3600
    LoginResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        status:
          type: string
          example: success
        message:
          type: string
          example: The user has been successfully logged in
        data:
          type: object
          properties:
            user:
              $ref: '#/components/schemas/User'
            token:
              $ref: '#/components/schemas/Token'
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        status:
          type: string
          example: success
        message:
          type: string
          example: Operation completed successfully
        data:
          type: object
          example: {}
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        status:
          type: string
          example: error
        message:
          type: string
          example: Error description
        data:
          type: object
          properties:
            error_code:
              type: string
              example: ERROR_CODE
  responses:
...
    UserNotFound:
      description: User not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            status: error
            message: User not found
            data:
              error_code: USER_NOT_FOUND
    InvalidToken:
      description: Invalid or expired token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            status: error
            message: Token is invalid or expired
            data:
              error_code: INVALID_TOKEN
    InvalidResetCode:
      description: Invalid or expired password reset code
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            status: error
            message: The password reset code has expired or is invalid
            data:
              error_code: PASSWORD_RESET_CODE_INVALID
...
    TooManyRequests:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            status: error
            message: Too many requests. Please try again later.
            data:
              error_code: RATE_LIMIT_EXCEEDED